pyproject.toml
[tool.poetry]
name = "mlflow-server"
version = "0.1.0"
description = "A server for managing MLflow experiments and models."
authors = ["Mattia Matteini <mattia.matteini@unibo.it>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">= 3.10.0 < 3.15.0"
mlflow = {extras = ["auth"], version = "^3.6.0"}
boto3 = "^1.40.75"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
python-on-whales = "^0.79.0"
ruff = "^0.14.7"
poethepoet = "^0.38.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
test = "pytest"
ruff-check = "ruff check"
ruff-fix = "ruff check --fix"
ruff-format = "ruff format"
format-check = "ruff format --check"
compile = "python -m compileall tests"
static-checks = [
    "ruff-check",
]
format = [
    "ruff-fix",
    "ruff-format"
]
